Fix flicker when transitioning to next slide
Immediately show the next slide but still fade the previous slide out. This way the next slide will be positioned behind the previous slide, and thus there will be no white flicker as the opacity of both slides drops below 1 (and thus the sum of the opacities is also below 1). Now it looks more smooth, especially at the header and footer of the slides.