Fix sharp transitions due to Z index and show()
The show() function produces a sharp transition. The intent was that the new slide would be behind the old one and the old slide would fade out slowly. However the new slide was sometimes before the old one so even if the old one faded out slowly you couldn't see it because the new one had already snapped to 100% opacity before it.